The analyse performance binary was updated to increase the number of bins used in the histrogram(s) for producing the jet energy resolution by a factor of 10.  This means that small flucturations in the events have a reduced impact on the jet eenrgy resolution metric.
The calirbation helper was edited so that it produced a histogram of the SimCalorimeterHit energy distribution depsoited in the ECal for single muon events.  This is needed so that the MIP scale can be set in the digitiser.  This is needed for correct use of the realistic ECal functionality in ILDCaloDigi. 

A similar distribution already existed in the code for calibration of the HCal ring and this is used for the calibration of the realistic HCal in ILDCaloDigi and so this did not require modification.